Fix bugs in accept-reject sampler for rtruncnorm().
Bug originally reported by Sune Karlsson with a partial fix. Further
investigation showed that more edge cases needed to be changed in the
selection heuristic for the sampling strategy in case 'a' and 'b' are
finite.

Additionally some debugging aids were added to the C source code.
These are not active by default but can be turned on at compile time
by defining the DEBUG directive.

Finally the 'urs_a_b' sampler was marginally simplifed.